# src/App/Http/Controllers/Auth/LogoutController.php
<?php

namespace App\Http\Controllers\Auth;

use App\Http\Controllers\Controller;
use Illuminate\Http\JsonResponse;
use Illuminate\Http\Request;

/**
 * @OA\Post(
 *     path="/api/auth/logout",
 *     tags={"AUTH"},
 *     summary="User logout",
 *     security={{"bearer": {}}},
 *     @OA\Response(
 *           response="202",
 *           description="result",
 *           @OA\JsonContent(ref="#/components/schemas/SuccessResponse"),
 *      ),
 *     @OA\Response(
 *          response="500",
 *          description="result",
 *          @OA\JsonContent(ref="#/components/schemas/ErrorResponse"),
 *      )
 * )
 */
class LogoutController extends Controller
{
    /**
     * @param Request $request
     * @return JsonResponse
     */
    public function __invoke(Request $request): JsonResponse
    {
        $user = currentUser();

        $user->currentAccessToken()->delete();

        return $this->acceptedResponse('ok');
    }
}